Add tracking socketio disconnection
This patch sets to null the socket associated to the connected client.
When the backend detects that there isn't connected client (physical
device), the requested will be accepted without confirming the action
at the physcal device. This approach allows testing assistant initiated
requests to be completed interacting with the Firebase only, without
waiting for the RPi3.